Fix determining `FullProgramName()`
When the program name doesn't contain any slashes, that means it was
invoked on the command-line normally (as e.g. `rbenv`), and the system
found the executable in PATH. To determine full path to the executable,
we must search the PATH ourselves.

Then the program name contains even a single slash, that means it was
invoked either with an absolute path, or relatively to the current
directory. The `Abs()` method will take care of that as before.